Created $.valueview and sub-system for creating DataValue edit view widgets
This introduces $.fn.valueview as well as $.valueview.widget. The later serves as a widget factory, similar to $.widget but with the purpose to create widgets which will be used by $.fn.valueview and for a specific kind of DataValue.
Also created $.valueview.Widget which inherits from $.Widget and extends it with functions required for the views. This is not yet done, the definition will be enhanced in a follow-up.

* patch set 2: fixed various issues in jquery.valueview.js

Change-Id: I9285c50f74e26910360258b7a6f09165eacd5022