(bug 49264, bug 48965) validate time format.
This provides validation of the provided date/time string
when constructing a TimeValue object.

It does not however validate the calendar model, since the
DataValue does not (and should not) know which models would be
valid. This needs to be done by a Validator based on the
proiperty's data type (not the data value type).

Note that this contributes to the solution of the mentioned bugs,
but is not a complete fix.

Change-Id: I6990983ef0c0cad7c9d4f271bdf803902b94230b