Revert "(bug 49264, bug 48965) validate time format."
This reverts commit 7accc7bfacbec38ec44415d33df12b6d8f022ed5

Strict validation when constructing TimeValue objects will break a lot
of things if we find invalid data in the database (as is the case for
wikidata.org). 

When merging this, we need to at least have a clear plan to handle the
resulting failures gracefully when rendering, diffing, serializing, 
importing, etc.

Change-Id: I8916062971114d6bc3188dd4f0322c28cc54403e