(bug 48611) valueview expert test setups won't create instances anymore
Instead, the actual valueview instances will only created as part of the actual tests. This will
prevent from side-effects happening during expert's initialization. E.g. widgets being initialized,
causing global DOM nodes to be added to the body which would interfere with other tests assumptions
about the state of the DOM.

Enhanced qunit.parameterize to take an array of functions returning objects as test cases for this
to work conveniently.

Change-Id: Id3af61c498f4cd71e31b3424e71cce8d2eab4a83